timeline/2024/04/community/requirements.txt
AgentLogin==3.0.0
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
captcha==0.5.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2
django-ckeditor==6.7.1
django-haystack==3.2.1
django-js-asset==2.2.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
idna==3.7
itypes==1.2.0
jieba==0.42.1
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
mysqlclient
oauthlib==3.2.2
pika==1.3.2
pillow==10.3.0
prompt-toolkit==3.0.43
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Whoosh==2.7.4
py3Fdfs
mutagen
